+++ scenario/globbing_resolving/lib/test.cpp
@@ -1,0 +1,6 @@
+#include "test.hpp"
+
+int test()
+{
+    return 4;
+}