+++ scenario/globbing_resolving/lib/test.hpp
@@ -1,0 +1,6 @@
+#ifndef HEADER_test_hpp_ALREADY_INCLUDED
+#define HEADER_test_hpp_ALREADY_INCLUDED
+
+int test();
+
+#endif